requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Template engine
jinja2>=3.1.3

# OCR processing
pytesseract>=0.3.10
Pillow>=10.4.0

# Google Sheets integration
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Scheduling
APScheduler>=3.10.4

# Session management
itsdangerous>=2.1.2
starlette>=0.36.0

# Date parsing
python-dateutil>=2.8.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
ruff>=0.1.14

# YAML parsing for contract tests
pyyaml>=6.0.0
